Update path for MSCollectionViewCalendarLayout
So that it works with latest cocoa pods.

Previously when running ```pod install``` after cloning the repo it errored with this message

```
nalyzing dependencies
Fetching podspec for `MSCollectionViewCalendarLayout` from `../`
[!] Unable to satisfy the following requirements:
- `MSCollectionViewCalendarLayout (from `../`)` required by `Podfile`
```

So I did the edit above and it worked. Strangely though if I change it back to just ../ it works second time. But if I re clone the repo it fails as before.